requirements.txt
asgiref==3.4.1
asttokens==2.0.5
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
colorama==0.4.4
cryptography==3.4.8
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.6
django-allauth==0.45.0
djangorestframework==3.12.4
dnspython==1.16.0
executing==0.8.0
icecream==2.1.1
idna==3.2
oauthlib==3.1.1
Pillow==8.3.2
psycopg2==2.9.1
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.10.0
PyJWT==2.1.0
pymongo==3.12.0
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.2.4
toml==0.10.2
urllib3==1.26.7
